+++ experimental-examples/tina-cloud-starter/.tina/config.tsx
@@ -1,4 +1,4 @@
-import { defineConfig } from 'tinacms'
+import { defineConfig, tinaTableTemplate } from 'tinacms'
 import { contentBlockSchema } from '../components/blocks/content'
 import { featureBlockSchema } from '../components/blocks/features'
 import { heroBlockSchema } from '../components/blocks/hero'
@@ -85,6 +85,7 @@ const config = defineConfig({
             label: 'Body',
             name: '_body',
             templates: [
+              tinaTableTemplate,
               {
                 name: 'DateTime',
                 label: 'Date & Time',