+++ experimental-examples/tina-cloud-starter/components/posts/post.tsx
@@ -38,6 +38,8 @@ const components: Components<{
   }
 }> = {
   code_block: (props) => <Prism {...props} />,
+  th: (props) => <th className="bg-gray-100 font-bold" {...props} />,
+  td: (props) => <td className="bg-gray-100" {...props} />,
   BlockQuote: (props: {
     children: TinaMarkdownContent
     authorName: string