+++ packages/@tinacms/graphql/src/resolve.ts
@@ -13,12 +13,9 @@ import type { GraphQLResolveInfo } from 'graphql'
 import type { Database } from './database'
 import { NAMER } from './ast-builder'
 import { handleFetchErrorError } from './resolver/error'
-import {
-  checkPasswordHash,
-  generatePasswordHash,
-  mapUserFields,
-} from './auth/utils'
+import { checkPasswordHash, mapUserFields } from './auth/utils'
 import _ from 'lodash'
+import { NotFoundError } from './error'
 
 export const resolve = async ({
   config,
@@ -506,12 +503,16 @@ export const resolve = async ({
     if (res.errors) {
       if (!silenceErrors) {
         res.errors.map((e) => {
-          console.error(e.toString())
+          if (e instanceof NotFoundError) {
+            // do nothing
+          } else {
+            console.error(e.toString())
 
-          if (verboseValue) {
-            console.error('More error context below')
-            console.error(e.message)
-            console.error(e)
+            if (verboseValue) {
+              console.error('More error context below')
+              console.error(e.message)
+              console.error(e)
+            }
           }
         })
       }