+++ packages/tinacms/src/unifiedClient/index.ts
@@ -1,4 +1,6 @@
 import fetchPonyfill from 'fetch-ponyfill'
+import type { GraphQLError } from 'graphql'
+import type { Config } from '@tinacms/schema-tools'
 
 const { fetch: fetchPonyfillFN, Headers: HeadersPonyfill } = fetchPonyfill()
 
@@ -12,10 +14,12 @@ export interface TinaClientArgs<GenQueries = Record<string, unknown>> {
   url: string
   token?: string
   queries: (client: TinaClient<GenQueries>) => GenQueries
+  errorPolicy?: Config['client']['errorPolicy']
 }
 export type TinaClientRequestArgs = {
   variables?: Record<string, any>
   query: string
+  errorPolicy?: 'throw' | 'include'
 } & Partial<Omit<TinaClientArgs, 'queries'>>
 
 export type TinaClientURLParts = {
@@ -27,20 +31,25 @@ export type TinaClientURLParts = {
 export class TinaClient<GenQueries> {
   public apiUrl: string
   public readonlyToken?: string
-  /**
-   *
-   */
   public queries: GenQueries
-  constructor({ token, url, queries }: TinaClientArgs<GenQueries>) {
+  public errorPolicy: Config['client']['errorPolicy']
+  constructor({
+    token,
+    url,
+    queries,
+    errorPolicy,
+  }: TinaClientArgs<GenQueries>) {
     this.apiUrl = url
     this.readonlyToken = token?.trim()
     this.queries = queries(this)
+    this.errorPolicy = errorPolicy || 'throw'
   }
 
-  public async request<DataType extends Record<string, any> = any>(
-    args: TinaClientRequestArgs
-  ): Promise<{ data: DataType; query: string }> {
-    const data: DataType = {} as DataType
+  public async request<DataType extends Record<string, any> = any>({
+    errorPolicy,
+    ...args
+  }: TinaClientRequestArgs) {
+    const errorPolicyDefined = errorPolicy || this.errorPolicy
     const headers = new HeadersDefined()
     if (this.readonlyToken) {
       headers.append('X-API-KEY', this.readonlyToken)
@@ -73,15 +82,15 @@ export class TinaClient<GenQueries> {
       )
     }
     const json = await res.json()
-    if (json.errors) {
+    if (json.errors && errorPolicyDefined === 'throw') {
       throw new Error(
         `Unable to fetch, please see our FAQ for more information: https://tina.io/docs/errors/faq/
-
         Errors: \n\t${json.errors.map((error) => error.message).join('\n')}`
       )
     }
     return {
       data: json?.data as DataType,
+      errors: (json?.errors || null) as GraphQLError[] | null,
       query: args.query,
     }
   }