+++ examples/tina-self-hosted-demo/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @tinacms/self-hosted-starter
 
+## 0.1.36
+
+### Patch Changes
+
+- Updated dependencies [b6fbab887]
+- Updated dependencies [4ae43fdde]
+- Updated dependencies [aec44a7dc]
+  - tinacms@1.5.22
+  - @tinacms/datalayer@1.2.27
+
+## 0.1.35
+
+### Patch Changes
+
+- Updated dependencies [177002715]
+- Updated dependencies [e69a3ef81]
+- Updated dependencies [c925786ef]
+- Updated dependencies [9f01550dd]
+  - tinacms@1.5.21
+  - @tinacms/datalayer@1.2.26
+
 ## 0.1.34
 
 ### Patch Changes