+++ experimental-examples/tina-cloud-starter/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @tinacms/starter
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies [b6fbab887]
+- Updated dependencies [4ae43fdde]
+- Updated dependencies [aec44a7dc]
+  - @tinacms/mdx@1.3.20
+  - tinacms@1.5.22
+  - @tinacms/vercel-previews@0.0.18
+
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [177002715]
+- Updated dependencies [e69a3ef81]
+- Updated dependencies [c925786ef]
+- Updated dependencies [9f01550dd]
+  - tinacms@1.5.21
+  - @tinacms/vercel-previews@0.0.17
+
 ## 0.1.16
 
 ### Patch Changes