+++ packages/@tinacms/datalayer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # tina-graphql
 
+## 1.2.27
+
+### Patch Changes
+
+- Updated dependencies [857414612]
+- Updated dependencies [a58c5b12f]
+- Updated dependencies [aec44a7dc]
+  - @tinacms/graphql@1.4.27
+
+## 1.2.26
+
+### Patch Changes
+
+- Updated dependencies [c244d963a]
+- Updated dependencies [3b214ec6b]
+- Updated dependencies [8bd85b15e]
+  - @tinacms/graphql@1.4.26
+
 ## 1.2.25
 
 ### Patch Changes