+++ packages/next-tinacms-cloudinary/CHANGELOG.md
@@ -1,5 +1,25 @@
 # next-tinacms-cloudinary
 
+## 4.3.7
+
+### Patch Changes
+
+- Updated dependencies [b6fbab887]
+- Updated dependencies [4ae43fdde]
+- Updated dependencies [aec44a7dc]
+  - tinacms@1.5.22
+
+## 4.3.6
+
+### Patch Changes
+
+- 177002715: Fix/media thumbnail
+- Updated dependencies [177002715]
+- Updated dependencies [e69a3ef81]
+- Updated dependencies [c925786ef]
+- Updated dependencies [9f01550dd]
+  - tinacms@1.5.21
+
 ## 4.3.5
 
 ### Patch Changes